pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "python-dotenv>=1.1.1",
]
